LPD-56403 Rename apiExplorerPage to fit convention
